+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Issue #36: Change motor controller to manage stiction
+- Issue #42: Merge filter and speedmeter
 
 ### Fixed